Ensure that DOM ids exist and move jsx file to new folder
