+++ kuma/javascript/src/document.jsx
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { css } from '@emotion/core';
 
-import A11yNav from './components/a11y-nav.jsx';
+import A11yNav from './a11y/a11y-nav.jsx';
 import Article from './article.jsx';
 import Banners from './banners.jsx';
 import { gettext } from './l10n.js';
