+++ kuma/javascript/src/landing-page.jsx
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react';
 
-import A11yNav from './components/a11y-nav.jsx';
+import A11yNav from './a11y/a11y-nav.jsx';
 import Banners from './banners.jsx';
 import GAProvider from './ga-provider.jsx';
 import Header from './header/header.jsx';
