+++ kuma/landing/jinja2/landing/react_homepage.html
@@ -30,7 +30,7 @@
                 None)|safe }}
 
 <!-- top search area -->
-<div class="home-masthead home-masthead-no-search">
+<div id="content" class="home-masthead home-masthead-no-search">
     <svg class="masthead-background" width="1340" height="426" xmlns="http://www.w3.org/2000/svg" aria-hidden="true"><g fill="#62ADCD" fill-rule="evenodd"><path d="M1097.974 131.552v17.364h-8.698v8.684h8.698v-8.684h8.702v-8.68h8.701v-8.684h-17.403zm0 26.048v17.364h17.403v-8.683h-8.701V157.6h-8.702zm34.802-26.048v8.683h8.701v8.681h8.699v-17.364h-17.4zm17.4 17.364v8.684h-8.699v8.68h-8.701v8.684h17.4V157.6h8.701v-8.684h-8.701zM954.187 42.69v8.684h-8.701v26.045h8.701V51.374h8.701V42.69h-8.701zm0 43.413h8.701v-8.684h-8.701v8.684zm26.099-34.73h8.702V42.69h-8.702v8.683zm8.702 0V77.42h-8.702v8.683h8.702V77.42h8.701V51.373h-8.701zM860.267 42.69v17.365h-8.698v8.683h-8.702v17.365h8.702V68.738h8.698v-8.683h8.705V42.69zm43.505 0v17.365h-8.704v8.683h-8.699v17.365h8.699V68.738h8.704v-8.683h8.699V42.69zm293.586-.102v17.475h-8.755v8.738h-8.753v17.476h8.753V68.8h8.755v-8.738h8.756V42.588zm43.778 0v17.475h-8.758v8.738h-8.753v17.476h8.753V68.8h8.758v-8.738h8.756V42.588zm43.776 0v17.475h-8.756v8.738h-8.753v17.476h8.753V68.8h8.756v-8.738h8.759V42.588zm43.778 0v17.475h-8.755v8.738h-8.753v17.476h8.753V68.8h8.755v-8.738h8.759V42.588zM515.054.638v8.943h-8.916v26.821h8.916V9.581h8.911V.638h-8.91zm0 44.707h8.911v-8.943h-8.91v8.943zM541.79 9.581h8.916V.64h-8.917v8.94zm8.916 0v26.822h-8.917v8.942h8.917v-8.942h8.91V9.58h-8.91zM181.974 146.552v17.364h-8.698v8.684h8.698v-8.684h8.702v-8.68h8.701v-8.684h-17.403zm0 26.048v17.364h17.403v-8.683h-8.701V172.6h-8.702zm34.802-26.048v8.683h8.701v8.681h8.699v-17.364h-17.4zm17.4 17.364v8.684h-8.699v8.68h-8.701v8.684h17.4V172.6h8.701v-8.684h-8.701zM38.187 57.69v8.684h-8.701v26.045h8.701V66.374h8.701V57.69h-8.701zm0 43.413h8.701v-8.684h-8.701v8.684zm26.099-34.73h8.702V57.69h-8.702v8.683zm8.702 0V92.42h-8.702v8.683h8.702V92.42h8.701V66.373h-8.701zm208.37-8.785v17.475h-8.755v8.738h-8.753v17.476h8.753V83.8h8.755v-8.738h8.756V57.588zm43.778 0v17.475h-8.758v8.738h-8.753v17.476h8.753V83.8h8.758v-8.738h8.756V57.588zm43.776 0v17.475h-8.756v8.738h-8.753v17.476h8.753V83.8h8.756v-8.738h8.759V57.588zm43.778 0v17.475h-8.755v8.738h-8.753v17.476h8.753V83.8h8.755v-8.738h8.759V57.588zm43.778 0v17.475h-8.758v8.738h-8.753v17.476h8.753V83.8h8.758v-8.738h8.756V57.588zM-24.224 367.552v8.683h8.701v8.681h8.699v-17.364h-17.4zm17.4 17.364v8.684h-8.699v8.68h-8.701v8.684h17.4V393.6h8.701v-8.684h-8.701zm47.182-106.328v17.475h-8.755v8.738H22.85v17.476h8.753V304.8h8.755v-8.738h8.756v-17.475zm43.778 0v17.475h-8.758v8.738h-8.753v17.476h8.753V304.8h8.758v-8.738h8.756v-17.475zm43.776 0v17.475h-8.756v8.738h-8.753v17.476h8.753V304.8h8.756v-8.738h8.759v-17.475zm43.778 0v17.475h-8.755v8.738h-8.753v17.476h8.753V304.8h8.755v-8.738h8.759v-17.475zm43.778 0v17.475h-8.758v8.738h-8.753v17.476h8.753V304.8h8.758v-8.738h8.756v-17.475z" opacity=".143"/><path d="M961.941 381.965v17.364h-8.699v8.684h8.7v-8.684h8.7v-8.68h8.702v-8.684h-17.403zm0 26.048v17.364h17.403v-8.683h-8.701v-8.681h-8.702zm34.801-26.048v8.683h8.702v8.681h8.699v-17.364h-17.4zm17.4 17.364v8.684h-8.698v8.68h-8.702v8.684h17.4v-17.364h8.702v-8.684h-8.701zM818.154 293.103v8.684h-8.702v26.046h8.702v-26.046h8.701v-8.684h-8.701zm0 43.413h8.701v-8.683h-8.701v8.683zm26.099-34.729h8.702v-8.684h-8.702v8.684zm8.702 0v26.045h-8.702v8.684h8.702v-8.684h8.701v-26.045h-8.701zm-128.721-8.684v17.365h-8.699v8.684h-8.701v17.364h8.701v-17.364h8.7v-8.684h8.703v-17.365zm43.505 0v17.365h-8.704v8.684h-8.7v17.364h8.7v-17.364h8.704v-8.684h8.699v-17.365zm293.586-.102v17.476h-8.756v8.737h-8.753v17.476h8.753v-17.476h8.756v-8.737h8.756V293zm43.778 0v17.476h-8.758v8.737h-8.753v17.476h8.753v-17.476h8.758v-8.737h8.756V293zm43.776 0v17.476h-8.756v8.737h-8.753v17.476h8.753v-17.476h8.756v-8.737h8.758V293zm43.778 0v17.476h-8.756v8.737h-8.753v17.476h8.753v-17.476h8.756v-8.737h8.759V293zm43.778 0v17.476h-8.758v8.737h-8.753v17.476h8.753v-17.476h8.758v-8.737h8.756V293zm59.719-154.898v8.684h-8.702v26.046h8.702v-26.046h8.701v-8.684h-8.701zm0 43.413h8.701v-8.683h-8.701v8.683zm26.099-34.729h8.702v-8.684h-8.702v8.684zm8.702 0v26.045h-8.702v8.684h8.702v-8.684h8.701v-26.045h-8.701zm-128.721-8.684v17.365h-8.699v8.684h-8.701v17.364h8.701v-17.364h8.7v-8.684h8.703v-17.365zm43.505 0v17.365h-8.704v8.684h-8.7v17.364h8.7v-17.364h8.704v-8.684h8.699v-17.365z" opacity=".14"/></g></svg>
     <div class="center">
         <h1 class="highlight highlight-spanned"><span class="highlight-span">{{ _('Resources for developers, by developers.') }}</span></h1>
